task3_image_captioning/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
Pillow>=8.3.0
numpy>=1.21.0
matplotlib>=3.4.0
requests>=2.25.0
